services/workflow-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-multipart>=0.0.6
aiofiles>=23.2.0
PyYAML>=6.0.0
openai>=1.12.0
msal>=1.28.0
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
redis>=5.0.0